refactor: Move indent function to separate file
